refactor(autocomplete): make triggerAutocomplete tag a transaction
