+++ packages/extensions/src/autocomplete/autocomplete.ts
@@ -1,5 +1,7 @@
 import { defineFacet, defineFacetPayload, pluginFacet, type Extension, type PluginPayload } from '@prosekit/core'
+import type { Transaction } from '@prosekit/pm/state'
 
+import { setTrMeta } from './autocomplete-helpers.ts'
 import { createAutocompletePlugin } from './autocomplete-plugin.ts'
 import type { AutocompleteRule } from './autocomplete-rule.ts'
 
@@ -13,6 +15,20 @@ import type { AutocompleteRule } from './autocomplete-rule.ts'
  */
 export function defineAutocomplete(rule: AutocompleteRule): Extension {
   return defineFacetPayload(autocompleteFacet, [rule])
+}
+
+/**
+ * Tags a transaction so that, when it is applied, autocomplete re-scans the text
+ * before the cursor and opens the menu if a rule matches. Returns the same
+ * transaction, so it can be chained.
+ *
+ * Autocomplete normally only opens while the user is typing. Use this to open a
+ * slash, mention, or tag menu imperatively, for example after inserting the
+ * trigger text in your own command, without dispatching a second transaction. It
+ * is intended for an empty (cursor) selection.
+ */
+export function triggerAutocomplete(tr: Transaction): Transaction {
+  return setTrMeta(tr, { type: 'scan' })
 }
 
 const autocompleteFacet = defineFacet<AutocompleteRule, PluginPayload>({
